[FIX BUG] caching

+++ src/main/java/com/example/demospringmongo/service/BookServiceImpl.java
@@ -1,9 +1,11 @@
 package com.example.demospringmongo.service;
 
+import com.example.demospringmongo.exception.EntityNotFoundException;
 import com.example.demospringmongo.model.Book;
 import com.example.demospringmongo.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
@@ -13,10 +15,12 @@ import java.util.Optional;
 @Service
 public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
+    private final CacheService cacheService;
 
     @Autowired
-    public BookServiceImpl(BookRepository bookRepository) {
+    public BookServiceImpl(BookRepository bookRepository, CacheService cacheService) {
         this.bookRepository = bookRepository;
+        this.cacheService = cacheService;
     }
 
     @Override
@@ -28,30 +32,46 @@ public class BookServiceImpl implements BookService {
     @Override
     @Cacheable(value = "books", key = "#id")
     public Optional<Book> getBookById(String id) {
-        return bookRepository.findById(id);
+        return Optional.ofNullable(bookRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("book not found with id: " + id)));
     }
 
     @Override
-    @CacheEvict(value = "books", allEntries = true)
+    @CachePut(value = "books", key = "#book.id")
     public Book saveBook(Book book) {
-        return bookRepository.save(book);
+        Book savedBook = bookRepository.save(book);
+//        cacheService.updateBookCache(savedBook);
+        return savedBook;
     }
 
     @Override
     @CacheEvict(value = "books", key = "#id")
     public Book updateBook(String id, Book bookDetails) {
-        Optional<Book> book = getBookById(id);
-        return book.map(b -> {
-            b.setTitle(bookDetails.getTitle());
-            b.setAuthor(bookDetails.getAuthor());
-            b.setPrice(bookDetails.getPrice());
-            return bookRepository.save(b);
-        }).orElse(null);
+        return bookRepository.findById(id)
+                .map(existingBook -> {
+                    updateBookDetails(existingBook, bookDetails);
+                    Book updatedBook = bookRepository.save(existingBook);
+                    cacheService.updateBookCache(updatedBook);
+                    return updatedBook;
+                }).orElseThrow(() -> new EntityNotFoundException("book not found with id: " + id));
     }
 
     @Override
     @CacheEvict(value = "books", key = "#id")
     public void deleteBook(String id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("book not found with id: " + id);
+        }
         bookRepository.deleteById(id);
+    }
+
+    private void updateBookDetails(Book book, Book bookDetails) {
+        if (bookDetails.getTitle() != null) {
+            book.setTitle(bookDetails.getTitle());
+        }
+        if (bookDetails.getAuthor() != null) {
+            book.setAuthor(bookDetails.getAuthor());
+        }
+        book.setPrice(bookDetails.getPrice());
     }
 }
